cleaning up armjoy
some cleanup and sample node cleaned up
