+++ src/rover_samples/rv_sample_package/CMakeLists.txt
@@ -10,14 +10,15 @@ find_package(ament_cmake REQUIRED)
 find_package(rclcpp REQUIRED)
 find_package(joy REQUIRED)
 find_package(std_msgs REQUIRED)
+find_package(rover_utils REQUIRED)
 # uncomment the following section in order to fill in
 # further dependencies manually.
 # find_package(<dependency> REQUIRED)
 include_directories(include/)
 #add_executable adds a 'target' for CMake to go in and do it's magic
 #The target's name here is 'sample_node', and while the naming doesn't need to match the source files, its good practice to keep them the same as this is what you use with ros2 run
-add_executable(sample_node src/sample_node.cpp include/sample_node.h)
-ament_target_dependencies(sample_node rclcpp joy)
+add_executable(sample_node src/sample_node.cpp include/sample_node.h include/sample_node_defines.h)
+ament_target_dependencies(sample_node rclcpp joy rover_utils)
